rename common_ui_targets _common_ui_targets
