+++ traitsui/testing/tester/qt4/implementation/check_list_editor.py
@@ -11,7 +11,7 @@
 
 from traitsui.qt4.check_list_editor import CustomEditor
 from traitsui.testing.tester import command, locator
-from traitsui.testing.tester.common_ui_targets import _BaseSourceWithLocation
+from traitsui.testing.tester._common_ui_targets import _BaseSourceWithLocation
 from traitsui.testing.tester.editors._layout import column_major_to_row_major
 from traitsui.testing.tester.qt4 import _interaction_helpers
